% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{addcode}
\alias{addcode}
\title{addcode}
\usage{
addcode(code, code.add)
}
\arguments{
\item{code}{an existing code}

\item{code.add}{the code to add}
}
\value{
merge of the two code
}
\description{
add a code event to an another
}
