% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-Host.R
\name{addHost}
\alias{addHost}
\title{addHost}
\usage{
addHost(lhost, id)
}
\arguments{
\item{lhost}{a hostSet Object}

\item{id}{a character of host ID}
}
\value{
a HostSet of host object with there ID
}
\description{
add an Host to a HostSet
}
\examples{
lhost <- list()
lhost <- addHost(lhost,"42")

}
