% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{addIndex}
\alias{addIndex}
\title{addIndex}
\usage{
addIndex(index, id_host, time, code)
}
\arguments{
\item{index}{an index}

\item{id_host}{an host index in HostSet}

\item{time}{a time}

\item{code}{an event code}
}
\value{
the index updated (add a row or update one)
}
\description{
add to an index a new eventcode
}
