% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-Host.R
\name{addViralObs}
\alias{addViralObs}
\title{addViralObs}
\usage{
addViralObs(lhost, lvpop)
}
\arguments{
\item{lhost}{a HostSet}

\item{lvpop}{a ViralPopSet}
}
\value{
lhost update with viral population observed
}
\description{
load Viral pop observation in Host object
}
