% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-ViralPop.R
\name{alleleCount}
\alias{alleleCount}
\title{alleleCount}
\usage{
alleleCount(mat, seq.char = c("A", "T", "G", "C"))
}
\arguments{
\item{mat}{a genomique seq list as matrix by row}

\item{seq.char}{allele alphabet}
}
\value{
a matrix, each row as a unique seq and col as allele count by position
}
\description{
count allele at each position
}
