% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-ViralPop.R
\name{concatViralPop}
\alias{concatViralPop}
\title{concatViralPop}
\usage{
concatViralPop(lvpop, lid)
}
\arguments{
\item{lvpop}{a ViralPop Set}

\item{lid}{vector of viralpop id to concat}
}
\value{
a ViralPop object with ID concatenation from all IDs and time at 0.
}
\description{
concat several Viral population in one ViralPop object
}
