% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-Host.R
\name{createHost}
\alias{createHost}
\title{createHost}
\usage{
createHost(list_host)
}
\arguments{
\item{list_host}{a character vector of host ID}
}
\value{
a HostSet of host object with there ID
}
\description{
create a list of Host class object
}
\examples{
lh <- seq(1,30,1)
lhost <- createHost(lh)

}
