% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{createIndex}
\alias{createIndex}
\title{createIndex}
\usage{
createIndex(hostlist)
}
\arguments{
\item{hostlist}{a Hostset}
}
\value{
a data.frame with TIME, ID_HOST and EVENTCODE as columns
}
\description{
create an index of time id_host and event code
}
