% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-ViralPop.R
\name{diversity.pDistance}
\alias{diversity.pDistance}
\title{diversity.pDistance}
\usage{
diversity.pDistance(vpop)
}
\arguments{
\item{vpop}{a ViralPop object}
}
\value{
result
}
\description{
diversity calculation using Mean Pairwise Distance
}
