% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-ViralPop.R
\name{diversity.sfs}
\alias{diversity.sfs}
\title{diversity.sfs}
\usage{
diversity.sfs(vpop)
}
\arguments{
\item{vpop}{a viralPop class}
}
\value{
the site frequency spectra
}
\description{
Allele frequency spectrum or Site frequency spectra : the distribution of alternative allele frequencies across all sites of genetic sequences
}
