% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-time.R
\name{getDate}
\alias{getDate}
\title{getDate}
\usage{
getDate(time, format = "\%Y-\%m-\%dT\%H:\%M:\%S")
}
\arguments{
\item{time}{a timestamp or vector of}

\item{format}{Date format output (default \%Y-\%m-\%dT\%H:\%M:\%S)}
}
\value{
time as string date
}
\description{
Converte timestamp to Date (string)
}
