% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-Host.R
\name{getDiversity.pDistance}
\alias{getDiversity.pDistance}
\title{getDiversity.pDistance}
\usage{
getDiversity.pDistance(host, lvpop)
}
\arguments{
\item{host}{an Host object}

\item{lvpop}{a ViralPopSet object}
}
\value{
a data.frame with col as time of observation and p_distance
}
\description{
get pairwise distance of an host over viral population observated
}
