% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-Host.R
\name{getDiversity.sfs}
\alias{getDiversity.sfs}
\title{getDiversity.sfs}
\usage{
getDiversity.sfs(host, lvpop)
}
\arguments{
\item{host}{an Host object}

\item{lvpop}{an ViralPopSet object}
}
\value{
a list indexed by time that contains allele.time and count
}
\description{
get Allele Frequency Spectrum or Site Frequency spectra for observated viral pop of an host
}
