% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{getInfosByHostAndTime}
\alias{getInfosByHostAndTime}
\title{getInfosByHostAndTime}
\usage{
getInfosByHostAndTime(index, lhost)
}
\arguments{
\item{index}{an index}

\item{lhost}{a hosts list}
}
\value{
a data.frame with colnames (id, time, infectedby, status, probabilities, X ,Y)
}
\description{
get hosts informations, status, infectedby, coordinates and time
}
