% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-Host.R
\name{getStates}
\alias{getStates}
\title{getStates}
\usage{
getStates(lhost, id = NA)
}
\arguments{
\item{lhost}{a HostSet}

\item{id}{a vector of host id (default NA : all lhost)}
}
\value{
a data.frame
}
\description{
get Host(s) states
}
