% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-Host.R
\name{getTimeLine}
\alias{getTimeLine}
\title{getTimeLine}
\usage{
getTimeLine(lhost, id)
}
\arguments{
\item{lhost}{a hostSet}

\item{id}{a host ID}
}
\value{
a data.frame
}
\description{
get the time line of an host
}
