% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-time.R
\name{getTimestamp}
\alias{getTimestamp}
\title{getTimestamp}
\usage{
getTimestamp(date, format = "\%Y-\%m-\%dT\%H:\%M:\%S")
}
\arguments{
\item{date}{a date (as string) or vector of}

\item{format}{the date format (default \%Y-\%m-\%dT\%H:\%M:\%S)}
}
\value{
timestamp of the date(s)
}
\description{
Get the timestamp of Date
}
