% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-Host.R
\name{getTransmissionTree}
\alias{getTransmissionTree}
\title{getTransmissionTree}
\usage{
getTransmissionTree(lhost, id = NA)
}
\arguments{
\item{lhost}{a hostSet}

\item{id}{a vector of hosts ids (default NA : all host)}
}
\value{
a data.frame as source|target|time in columns
}
\description{
get a transmission tree as a data.frame
}
\examples{
path = system.file("extdata", "data-simul/", package="SMITIDstruct")
lhost <- list()
lhost <- loadTree(lhost,paste(path,"/tree.txt",sep=''))
print(getTransmissionTree(lhost))
}
