% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Host.R
\docType{class}
\name{Host}
\alias{Host}
\alias{Host-class}
\title{Class Host}
\description{
Spatio-temporal information about Host.
}
\details{
Object can be created by calling ...

rdname Host-class
}
\section{Slots}{

\describe{
\item{\code{ID}}{Host identifier}

\item{\code{coordinates}}{Host coordinates in time (as sf)}

\item{\code{states}}{Host States/Status (dob, Inf...)}

\item{\code{sources}}{data.frame of time and host id who infected this host}

\item{\code{offsprings}}{data.frame of time and host id who has been contamined by this host}

\item{\code{ID_V_POP}}{data.frame of time and index of Viral population Observation}

\item{\code{covariates}}{data.frame of time, cavariate and value of this host.}
}}

