% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-time.R
\name{is.juliendate}
\alias{is.juliendate}
\title{is.juliendate}
\usage{
is.juliendate(time)
}
\arguments{
\item{time}{a numeric}
}
\value{
TRUE if time is a julien day, otherwise FALSE
}
\description{
Chekc if a numeric is not a timestamp
}
