% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-time.R
\name{is.StringDate}
\alias{is.StringDate}
\title{is.StringDate}
\usage{
is.StringDate(date)
}
\arguments{
\item{date}{a string or a vector of string (without NA)}
}
\value{
TRUE if date contains date format
}
\description{
Check if a string represent a date
}
