% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-time.R
\name{is.timestamp}
\alias{is.timestamp}
\title{is.timestamp}
\usage{
is.timestamp(time)
}
\arguments{
\item{time}{a numeric}
}
\value{
TRUE if time >= 1971
}
\description{
Check if a numeric represent a timestamp
}
