% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{isInCode}
\alias{isInCode}
\title{isInCode}
\usage{
isInCode(code, thecode)
}
\arguments{
\item{code}{list of code to test}

\item{thecode}{the real code}
}
\value{
TRUE if code contain thecode otherwise FLASE
}
\description{
check a code contains a specific code
}
