% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-Host.R
\name{loadCoords}
\alias{loadCoords}
\title{loadCoords}
\usage{
loadCoords(lhost, dfCoords, id = "ID")
}
\arguments{
\item{lhost}{a HostSet}

\item{dfCoords}{a data.frame with host ID, time and longitude latitude values}

\item{id}{colname for host ID}
}
\value{
lhost updated
}
\description{
Load Hosts states
}
\examples{
path = system.file("extdata", "data-simul/", package="SMITIDstruct")
lhost <- list()
lhost <- loadTree(lhost,paste(path,"/tree.txt",sep=''))
coords <- read.table(file=paste(path,"/hosts_coords.txt",sep=''), header=TRUE, check.names=FALSE)
lhost <- loadCoords(lhost,coords)
}
