% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-Host.R
\name{loadCovs}
\alias{loadCovs}
\title{loadCovs}
\usage{
loadCovs(lhost, dfCovs, id = "ID", colCovs)
}
\arguments{
\item{lhost}{a HostSet}

\item{dfCovs}{a data.frame with host ID in rows and covariates in columns}

\item{id}{colname for host ID}

\item{colCovs}{colnames of covariates columns}
}
\value{
lhost updated with covariates
}
\description{
Load Hosts covariates
}
