% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-Host.R
\name{loadHost}
\alias{loadHost}
\title{loadHost}
\usage{
loadHost(file = "host.txt")
}
\arguments{
\item{file}{a file containing hosts data}
}
\value{
a list of Host object (HostSet)
include Class-Host.R
}
\description{
load host object from a file
}
