% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-Host.R
\name{loadTree}
\alias{loadTree}
\title{loadTree}
\usage{
loadTree(lhost = list(), file = "tree.txt", source = "ID-source",
  receptor = "ID-receptor", tinf = "Tinf", weight = "Weight")
}
\arguments{
\item{lhost}{a HostSet}

\item{file}{a file containing tree data}

\item{source}{column name for source ID}

\item{receptor}{column name for receptor ID}

\item{tinf}{column name for infection Time}

\item{weight}{column name of infection weight}
}
\value{
the lhost param update with sources and offsprings
}
\description{
load sources and offsprings from file
}
\examples{
path = system.file("extdata", "data-simul/", package="SMITIDstruct")
lhost <- list()
class(lhost) <- "hostSet"
lhost <- loadTree(lhost,paste(path,"/tree.txt",sep=''))
}
