% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-Host.R
\name{loadTreeDF}
\alias{loadTreeDF}
\title{loadTreeDF}
\usage{
loadTreeDF(lhost = list(), df = data.frame(), source = "ID-source",
  receptor = "ID-receptor", tinf = "Tinf", weight = "Weight")
}
\arguments{
\item{lhost}{a HostSet}

\item{df}{a data.frame containing tree data}

\item{source}{column name for source ID}

\item{receptor}{column name for receptor ID}

\item{tinf}{column name for infection Time}

\item{weight}{infection links probability}
}
\value{
the lhost param update with sources and offsprings
}
\description{
load sources and offsprings from a data.frame
}
