% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-ViralPop.R
\name{loadViralObs}
\alias{loadViralObs}
\title{loadViralObs}
\usage{
loadViralObs(id, time, file)
}
\arguments{
\item{id}{host pathogen ID}

\item{time}{time of the observation (numeric or Date)}

\item{file}{a fasta file}
}
\value{
a new ViralPop object
}
\description{
load a ViralPop object
}
