% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-ViralPop.R
\name{loadViralPopSet}
\alias{loadViralPopSet}
\title{loadViralPopSet}
\usage{
loadViralPopSet(lvpop = list(), list)
}
\arguments{
\item{lvpop}{a viralPopSet (default new one)}

\item{list}{a list (see details)}
}
\description{
load a list of viral populations
}
\details{
The list have to be on this format:  
list$HOST_ID$TIME$list$seq_id
                      $seq
                      $prop
A list indexed by host ID, follow by a list indexed by time (of observation).
The last list contains an array of seq_ID (sequence ID), an array of seq (sequence as characters), and an array of the count of seq.
example : 
$'HOST_42'$'2014-01-01T00:00:00'$seq_ID ["SEQ_1","SEQ_2"]
$'HOST_42'$'2014-01-01T00:00:00'$seq ["ACGT","TGCA"]
$'HOST_42'$'2014-01-01T00:00:00'$seq_ID ["46","6"]
}
