% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{plotDiversity.pDistance}
\alias{plotDiversity.pDistance}
\title{plotDiversity.pDistance}
\usage{
plotDiversity.pDistance(host, lvpop)
}
\arguments{
\item{host}{an Host object}

\item{lvpop}{a ViralPopSet object}
}
\description{
plot Mean Pairwise Distance for an host viralpop over time
}
