% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{plotDiversity.sfs}
\alias{plotDiversity.sfs}
\title{plotDiversity.sfs}
\usage{
plotDiversity.sfs(host, lvpop)
}
\arguments{
\item{host}{an Host object}

\item{lvpop}{an ViralPopSet object}
}
\description{
plot Allele frequency spetrum for an host viralpop over time
}
