% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-Host.R
\name{setStates}
\alias{setStates}
\title{setStates}
\usage{
setStates(lhost, dfStates, colStates = c(id = "ID", time = "time", states
  = "value"))
}
\arguments{
\item{lhost}{a HostSet}

\item{dfStates}{a data.frame with host ID and states and time in columns}

\item{colStates}{vector of the columns name, id, time and states}
}
\value{
the HostSet updated
}
\description{
set hosts states from a data.frame
}
