% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods-Host.R
\name{simulateStates}
\alias{simulateStates}
\title{simulateStates}
\usage{
simulateStates(lhost)
}
\arguments{
\item{lhost}{a HostSet}
}
\value{
lhost update with states from sources time ~
}
\description{
simulate states from sources infection
}
