% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMITIDstruct.R
\docType{package}
\encoding{UTF-8}
\name{SMITIDstruct-package}
\alias{SMITIDstruct-package}
\alias{_PACKAGE}
\alias{SMITIDstruct}
\title{Data Structure and Manipulation Tool for Host and Viral Population}
\description{
Statistical Methods for Inferring Transmissions of Infectious Diseases from deep sequencing data (SMITID).
 It allow sequence-space-time host and viral population data storage, indexation and querying.
}
\details{
\tabular{ll}{
         Package: \tab SMITIDstruct\cr
         Type: \tab Package\cr
         Version: \tab 0.0.5\cr
         Date: \tab 2019-06-14\cr
         License: \tab GPL (>=2)\cr
         }

The SMITIDstruct package contains functions and methods for manipulating Host and Viral population genotico-space-time data.
}
\examples{
\donttest{
## Run a simulation
library("SMITIDstruct")
demo.SMITIDstruct.run() 
}

}
\seealso{
\code{\link{demo.SMITIDstruct.run}}
}
\author{
Jean-Francois Rey \email{jean-francois.rey@inra.fr}

Maintainer: Jean-Francois Rey \email{jean-francois.rey@inra.fr}
}
