% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-ViralPop.R
\docType{class}
\name{ViralPop}
\alias{ViralPop}
\alias{ViralPop-class}
\title{Class ViralPop}
\description{
Viral population data containing genotypes
}
\section{Slots}{

\describe{
\item{\code{ID}}{Host identifier}

\item{\code{time}}{Observation time as numeric since 1970/01/01}

\item{\code{size}}{Qt of variants}

\item{\code{names}}{list of variants id with same sequence}

\item{\code{genotypes}}{all variants genotypes (as DNAStringSet)}

\item{\code{proportions}}{proportions of each variants}
}}

